Adding inital role
This is a feature complete role, but not ready for release (documentation, cleanup, etc.). Adding the role now so CI can be tested. The current README is in very rough/draft form!

Role as it stands supports configuring firewall services (but not ports) on CentOS/firewalld and Ubuntu/UFW. Testing this role has proved rather difficult in terms of functional tests so are currently disabled.